packaging: allow interactive NFS exports cleanup with engine-cleanup
When running engine-cleanup the user will be asked for removing any
configuration made by engine-setup in /etc/exports and for removing
the directories that were previously exported.

When running in unattended mode use the -n and -e options for
deciding what to do about the NFS exports.

If any change is made to /etc/exportfs, the NFS exports will be
refreshed and the file will be backed up.

Change-Id: I0548ab358d0cbed32ceff3ceacefc57f1c068df4
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=904149
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>