core: start 'prestarted VMs' immediately after update
This patch solves a bug in which after increasing the number of
prestarted VMs of a pool, it takes a while (which is defined by the
configuration value "VmPoolMonitorIntervalInMinutes") to start
more VMs to meet the updated number of prestarted VMs.

The solution is to trigger the pool monitoring job to run immediately,
after UpdateVmPoolWithVmsCommand updates the new properties of the
pool.

Change-Id: I636f496b241c042962cfec10880fbee00ce2bab4
Bug-Url: https://bugzilla.redhat.com/888396
Signed-off-by: Arik Hadas <ahadas@redhat.com>