core: Remove domains member from storage_pool
Removed the private member "domains" from the storage_pool class. This
member was never updated (in fact, there are no methods to modify it!),
and never queried (no such methods exists either), and was only used in
the equals(Object) and hashCode() methods. Since its value is always
null, it's useless to continue holding it.

Change-Id: Iee2b595457d14d8b5c7d6f1697d4dca1236fd1a5
Signed-off-by: Allon Mureinik <amureini@redhat.com>