core: Remove vdsGroups member from storage_pool
Removed the private member "vdsGroups" from the storage_pool class.
This member was never updated (in fact, there are no methods to
modify it!), and never queried (no such methods exists either), and
was only used in the equals(Object) and hashCode() methods.
Since its value is always null, it's useless to continue holding it.

Change-Id: I5ffad5811b390fd25562e677c7d852fb24ac4284
Signed-off-by: Allon Mureinik <amureini@redhat.com>