core: Remove vm_last_up_time member from VmDynamic
Removed the private member "vm_last_up_time" from the VmDynamic class.
This member was never updated (in fact, there are no methods to modify
it!), and never queried (no such methods exists either), and was only
used in the equals(Object) and hashCode() methods.
Since its value is always null, it's useless to continue holding it.

Change-Id: Iab5358b836c3cb7dd351bfeb25f1a197e534c1d1
Signed-off-by: Allon Mureinik <amureini@redhat.com>