core: Don't check equality in BE setters
Some old business entities used the change-check idiom in their setters:

public void setXXX (Object newValue) {
    if (!newValue.equals(dataMember) {
        dataMember = newValue;
    }
}

This was probably produced by the C# conversion, but is completely
useless in Java - in any event, at the end of the setXXX method the
dataMember's value will be equal to the one passed, and the equality
check is more expensive than just assigning the new value.

This patch removes this old idiom from the code.
As a nice side-bonus, it also removes a couple of StringHelper usages
from the code, as part of the ongoing effort to get rid of this class.

Change-Id: Iafac9ca646c6c988470bf49d0596d973531c4616
Signed-off-by: Allon Mureinik <amureini@redhat.com>