core: Don't use EqOp on constants in commons
This patch removes the usage on StringHelper.EqOp for string constants
from the commons package, and replaces them with the standard JDK string
equality checks such as equals and equalsIgnoreCase.

Note that the EqOp is just redundant in these cases, since a string
constant is known not to be null.

Change-Id: Id0936f213025c76b8c34df6d447e8067500a9887
Signed-off-by: Allon Mureinik <amureini@redhat.com>