engine: Disable global transaction at ExtendSANStorageDomain
During investigation on some bug, I noticed that we have too many transaction time out
exceptions on ExtendSANStorageDomainCommand.
The problem is that following command is trying to open a global transaction, this is a mistake,
transaction is a global resource and should be kept as short as possible.
The change is:
1. Disable global transaction
2. Start to use compensation in order to keep flow correct

Change-Id: I78c4c43c259a58c746caa0402bf5beb004e9a8c2
Signed-off-by: Michael Kublin <mkublin@redhat.com>