engine: slave interfaces are being overridden
When the client provides a bond with its slaves, the slaves are
being reloaded from the engine, causing the information not provided
by the client to be populated by the engine's nics data.

In order to unset previously configured slave configuration, the client
need to specifically override it so the unsupported slave's configuration
will be unset, if the slave was previously configured with any of the
following properties: network name, vlan notation, boot protocol, IP address,
gateway or netmask.

An attempt to provide a slave which isn't configured properly will be blocked by
the engine with a can-do-action message.

Change-Id: I4f38bc399059ad19e93fcad0f360957636ac7045
Bug-Url: https://bugzilla.redhat.com/907240
Signed-off-by: Moti Asayag <masayag@redhat.com>