core: Remove VdcActionParameterBase.httpSessionId
httpSessionId in VdcActionParameterBase is a private field that isn't
initialized in the constructor and has no set method. I.e., it is always
null.

Since it's actually a constant, there is no reason to hold it. This
patch removes the field and all its references. That refactoring, in
turn allows cleaning up login and logout flows a bit, as they have code
that only fires when this member is not null.

Change-Id: I239d13175a511291def1eb31ce55be17f376bf42
Signed-off-by: Allon Mureinik <amureini@redhat.com>