core: Non-plugin automatic invocation of console session
This patch adds a native way of connecting to a (SPICE and VNC) console from
WA/UP using external viewer (virt-viewer) installed on client system. It is
achieved by implementing a servlet that generates a config file (with all
needed information to connect) for virt-viewer that is served to the client
afterwards.

The old way of invoking console viewer (browser plugin) is still
kept, the configuration is set in the database.

The behavior for SPICE is driven by configuration. There are three modes
of behavior:
 - Plugin - The "old" way - spice-xpi is used.
 - Native - The native client is used.
 - Auto - If plugin is installed, it is used. Otherwise "native" implementation
   is used.

As for VNC - it is set to use the native implementation by default (i.e. there
is no VNC connection info popup dialog).

Backend changes:
 - Added servlets for serving console configuration.
 - Added ClientConsoleMode to vdc_options (default setting is 'Auto') and
   modified related enums.

Change-Id: I18b9d2efad15f5e9bb98a8176112b721c810205f
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=843410