frontend: Non-plugin automatic invocation of console session
This patch adds a native way of connecting to a (SPICE and VNC) console from
WA/UP using external viewer (virt-viewer) installed on client system. It is
achieved by implementing a servlet that generates a config file (with all
needed information to connect) for virt-viewer that is served to the client
afterwards.

The old way of invoking console viewer (browser plugin) is still
kept, the configuration is set in the database.

The behavior for SPICE is driven by configuration. There are three modes
of behavior:
 - Plugin - The "old" way - spice-xpi is used.
 - Native - The native client is used.
 - Auto - If plugin is installed, it is used. Otherwise "native" implementation
   is used.

As for VNC - it is set to use the native implementation by default (i.e. there
is no VNC connection info popup dialog).

Frontend changes:
 - removed classes that deal with displaying VNC connection info popup
 - added SpiceNativeImpl (implements ISpice) for invoking request for
   spice client configuration file
 - renamed SpiceInterfaceImpl to SpicePluginImpl
 - added marking interfaces to these two implementations
 - added class AbstractSpice that contains common properties of ISpice
   implementations
 - added ClientConsoleMode enum for representing console mode configuration
   (Auto, Native, Plugin)

Change-Id: Ib31e870deb7ecb3dac4cff25e49a3ebca4706a25
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=843410