engine: Added infrastructure for providing a canDoactions in case of locks
The following patch is introducing an infrastracture for providing a more convinient messages in case
of using lock util.
What is done: from now it is possible to pass appropriate message for every acquired lock.
What is not done: I did not change a generic message, so it still will be VdcBllMessages.ACTION_TYPE_FAILED_OBJECT_LOCKED,
but I provide an easy way to change it

Change-Id: I87698da36ddb6e7556e98612d299da68f2eb53dc
Signed-off-by: Michael Kublin <mkublin@redhat.com>