core: Removed unnecessary calls to StringFormat & StringHelper
Replaced calls to StringFormat.format to String.format where possible
(i.e. not code that compiles to GWT) and to simple String concatenation where
possible.
Also replaced calls to StringHelper.EqOp calls in
String.equals calls in cases that at least one of the args was not null.

Added @Override annotation to methods in the edited classes where needed

Change-Id: I57dc0d9d30f9e5cd4f0beacc8bfd35fe5023afc1
Signed-off-by: Tal Nisan <tnisan@redhat.com>