core: fix redundant storage server conn in db
Fix several problems related to "orphan" storage server connections
entries in DB.

The redundant entries were inserted when AddStorageServerConnectionCommand
succeeded, but another command that ran after it failed. Adding a new
command that in case of failure deletes the connection entry from DB fixes the
problem.

As part of the fix, added assignment of the connection id in UI in
StorageListModel after it is created, because otherwise it
would not know which connection to delete in case of a failure.

The fix takes care of use cases of NFS, POSIX and local storage
operations originated from UI.

Change-Id: Iea5468371514bd2c7bc043a6c5520e2864a09fe8
Bug-Url: https://bugzilla.redhat.com/815083
Signed-off-by: Alissa Bonas <abonas@redhat.com>