tools: error in engine-notifierd init script
When starting the notification service, the init script calls
/usr/share/ovirt-engine/bin/engine-notifier.sh ($NOTIFIER_SCRIPT)
through the daemon() function, which can run something as another
user--in this case, user 'ovirt'.

If the init script is run from a directory that 'ovirt' doesn't have
permission to access, a `find` in engine-notifier.sh will print an
error.  Prevent this error by having the init script change to the
ovirt home directory before starting the daemon.

Change-Id: I83622cbb8b39ea5792907e6efc68d824ca4b02da
Bug-Url: https://bugzilla.redhat.com/885446
Signed-off-by: Greg Padgett <gpadgett@redhat.com>