userportal: wrong display type on run once
This patch solves a bug that when running a VM using 'run once' mode
from the user-portal, the VM is starting with its default display type
even though other display type was selected for it from the 'run once'
dialog.

This patch also fix the default selection of the display type in the
'run once' dialog of user-portal to be the default display type of the
VM (there was no default selection of display type in the dialog
before).

Change-Id: I665ab4578fea13fce6a22b77ee0c5d627b5c235b
Bug-Url: https://bugzilla.redhat.com/907776
Signed-off-by: Arik Hadas <ahadas@redhat.com>