engine: Fixing compensation for RemoveVmTemplateCommand and possible other commands
So, the fix is simple - change name.
After a restart of engine - RemoveVmTemplate command is created based on parameters,
but with old name vmTemplateId did not pass json serialization and get default value - Guid.Empty,
which is id of blank template.
After that , we are trying to delete blank template and failed (We are lucky) - means command failed, a origin template left
with status Locked

Change-Id: Ifb042d6cb8276a3c457a94823e90c1e6936cfa69
Signed-off-by: Michael Kublin <mkublin@redhat.com>