engine: Null Pointer Exception when during preview mode action, service ovirt-engine restart
The following bug was caused because of wrong member name, which  was caused by not passing
serialization by dstSnapshotId, and after restoring it from DB after restart its value was Guid.Empty,
so because of that we usually failed at endSuccessfully because of NullPointerException.
Thise caused for VM to be stuck in status locked.

Change-Id: I273369d20ae3db3e9598f33ff75eb094dd420f65
Bug-Url: https://bugzilla.redhat.com/871802
Signed-off-by: Michael Kublin <mkublin@redhat.com>