core: Add storage space validation to removeSnapshot
Today there is no validation in RemoveSnapshotCommand to ensure there is enough
space on the domain to merge the disk snapshots.
This space is needed on remove snapshot process since VDSM creates a new
temporary volume, as part of the merge process, which might be large as
the volume's actual size.

The proposed fix introduces a new validation in the command which
validates whether each storage domain has enough free space to perform removeSnapshot.

Change-Id: Ifbebc5af8e7489bccfa5d8e79695fe21c5371ec9
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>