webadmin: Gluster Hooks - conflict resolution popup
New Action 'Resolve Conflicts' is added to gluster hooks tab.
This action will be enabled if the hook is not in sync in the
hosts of the cluster and engine. Engine copy of the hook will
be considered as a Master copy.

On clicking 'Resolve Conflicts', a popup will be shown explaining
the reason for the conflict. The conflict could be any one/mutiple of
the following

1.MISMATCH IN CONTENT - hook content is different across hosts
2.MISMATCH IN STATUS - hook status is different across hosts
3.MISSING HOOK - hook is not present in some hosts

The user will be able to view the content and status of the hook from different
sources (hosts and Engine).

Finally the user can select the resole actions which will resolve the conflicts.

Change-Id: I10bc9c7dba7cb1be320bb0e12f3b7bf7a8beb485
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>