core: Change PM delay for Start/Stop operations
currently when a Start/Stop operation is invoked, the default number of
retries is 3 while the delay between retries is 60 sec.
That means that if when we initially check the status after a Start/Stop
operation it does not reach on/off status we will have to wait 60
seconds to sample the status again

A shorter delay of 10 seconds while keeping the total default waiting
time untouched (means modifying default number of retries to 18 will
improve user experience in PM operations

Default config values were changed to:

FenceStartStatusDelayBetweenRetriesInSec=10
FenceStartStatusRetries=18
FenceStopStatusDelayBetweenRetriesInSec=10
FenceStopStatusRetries=18

Since those values are not exposed to engine-config, we can simply
override the installed value.

Change-Id: I876d98013e2c3874680660940f78b8f14778fa92
Signed-off-by: Eli Mesika <emesika@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=908778
Signed-off-by: Eli Mesika <emesika@redhat.com>