core: Remove user_sessions table
The user_sessions table is part of the old login/logout mechanism,
which is no longer in use.

Before this patch, the table still existed, as well as code that
queries it and deletes from it, but no occurrences of inserting or
updating it, which means it was effectively useless.

This patch drops it from the database, and removes any remnants of code
that still reference it, including the session_count column of the
users table (and the corresponding members of the DbUser class).

Change-Id: I46952752950dcfeb4f917d532b80f15df7ff19c6
Signed-off-by: Allon Mureinik <amureini@redhat.com>