core: add ability edit NFS path in webadmin
Add the option to edit NFS path and the overriden
additional options of storage domain (retransmissions, timeout, version)
in webadmin UI --> Storage tab.

Editing these properties is available only when
the storage domain is in maintenance status and is a data/master domain,
and for storage of nfs type. These are the only properties which are
editable in this status.

Editing is still available for storage domains in status Active/Mixed, but
only for name and description - like it was before this patch.

Related-To:https://bugzilla.redhat.com/show_bug.cgi?id=835543
Change-Id: Ifaff5344ff191d6bdf53cc706c7bb796167a56b3
Signed-off-by: Alissa Bonas <abonas@redhat.com>