core: Make IVdcUser extend Serializable
Define IVdcUser as extends Serializeable so Serializable classes
code inspection (such as FindBugs) applied to classes containing it do
not fret about a Serializable class with a non Serializable member.

Note that this does not change the current behaviour, as VdcUser, the
only known implementation of IVdcUser in fact implements Seriazlizable.

Change-Id: Id02c1f62fe3406ee0bceaf75cb9d79ce034e4a4d
Signed-off-by: Allon Mureinik <amureini@redhat.com>