core: upgrade issue
Patch addresses the following problem
taskcleaner needs dbfunctions and dbcustomfunctions
taskcleaner and dbfunctions*.sh are in dbscripts rpm, which is
locked, and downloaded only AFTER we clean the tasks during upgrade.

The ONLY viable option is to separate taskcleaner and its SP installing
from dbscripts, and move it into the setup rpm, in that way, it will be
always available.
In order to do so, we need to do the following:
1. Make taskcleaner a standalone script (e.g, remove the dependency on
dbfunctions)
2. Move it outside of dbscripts folder (to tools/dbutils)
3. Alter the rpm spec file to include the standalone taskcleaner in
rhevm-setup rpm, and deploy it outside of dbscripts.
(will be done in a separate patch)

Change-Id: I82e9ccd3f621f1b524d4f7ed587fe9f266ffa8de
Signed-off-by: Eli Mesika <emesika@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=909097