tools: Don't send events on startup
Currently the event notifier sends all the pending events during
startup. This can flood the recipients and it is not very useful as old
events are not valuable. This patch adds a new DAYS_TO_SEND_ON_STARTUP
parameter to the notifier. The value of this parameter is the number of
days of old audit log messages that will be sent during startup.
Messages older than that will be marked as processed and not sent.

Change-Id: Id1b8044533d29fa69585a157d3f8409c81a437cd
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>