core: When updating entity name the event...
When updating entity name the event message should contain the
former entity name.

This patch adds a event log message when an entity is renamed so the
proceeding log message on the entity update that uses the new name is
clear.

This is done by introducing a RenamedEntityInfoProvider interface that
all relevant Update*Command are implementing.
The interface forces those commands to state the affected entity name,
the old entity name, the new entity name and to set the relevant object
ID such that the relevant event will be shown when the entity is
selected in the UI.
The CommandBase is checking for each command if it implements the
RenamedEntityInfoProvider interface and checks if the entity was
renamed. If it founds that the entity was renamed, a proper message is
written to the audit log.

Change-Id: I17f0049ca0ffb1a3868ba404b112a8c10ee7778e
Signed-off-by: Eli Mesika <emesika@redhat.com>