frontend: Use console popup dialog in webadmin
This patch extracts the console popup dialog from userportal code and moves it
to common base so that it can be used by webadmin as well.

It makes use of existing classes for console support (ConsoleUtils,
ConsoleManager, ConsolePopupView/PresenterWidget, ConsolePopupModel...)
These classes worked for UserPortalItemModel only. To be possible to use them
in webadmin too, a common interface was introduced - HasConsoleModel (this
provides access to the console related fields that are encapsulated in
corresponding models (UserPortalItemModel, VmListModel).

This patch also encapsulates cached console models from VmListModel and
IUserPortalListModel to the class ConsoleModelCache, that encapsulates map of
console model and operations that are common for VLM and IPULM and  were
duplicated.

Minor changes:
- Removed boolean flag hasAdditionalConsole from VmListModel and
UserPortalItemModel as this flag introduces only additional state that has to
be kept up to date. The additional console can be tested by comparing
additionalConsoleModel to null.

Change-Id: I00158bbd8bd488dc527821b5f088fe598c9c2713
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>