webadmin: Disable revomal of plugged vnic on a running vm
On Networks->Virtual Machines the remove action should be disabled if
the vm is running and the vnic is plugged.
Also, it should be disabled in all the vm states that are not UP or DOWN.
Virtual Machines->Network interfaces already works correctly.

Also a little code fix- there is no need to check the HotPlugEnabled
property when deciding if the user can remove the vnic.
It is enough to make sure the vnic is unplugged. If HotPlug is not enabled
the vnic is always plugged.

Change-Id: I795096615e8ef5118cceb56155aa6806af736ebb
Bug-Url: https://bugzilla.redhat.com/910021
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>