webadmin: validation for "Use Host CPU" setting
To enable CPU Host, one of the two no-migrate flags for a VM (run on
selected host, or allow migration only upon admin request) must be
enabled.  If CPU Host was disabled along with the no-migrate flag in
the same save action, the new CPU Host value was not propagated to the
VM object which caused a validation error.

This patch propagates the CPU Host value to the VM object, allowing
the CPU Host flag and no-migrate flag to be disabled in the same edit.

Change-Id: I786a725f6ff20e64d2d094b4c8c183c907978229
Bug-Url: https://bugzilla.redhat.com/905843
Signed-off-by: Greg Padgett <gpadgett@redhat.com>