userportal: quota permissions show extended view wrongfully
In the userportal we show extended view if the user has other
permissions associated with him that are not included in user role.
'Consume Quota' is a user role and action group that is not included
in user role, therefore, in client side, we will add a check for
consume quota action group, and ignore/skip it.

Change-Id: I435bf61277bb77fd4e400ee738b74f71024f57d3
Bug-Url: https://bugzilla.redhat.com/905144
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>