gluster: resolve hostname before checking ip in db
At two places, we are checking if a server with given ip address exists
in the engine DB. However it is possible that the value passed is a
hostname, and it's not same as what is present in the host_name field.
To handle such cases, modified the code to make sure that we resolve
the given ip/hostname, and check for the resolved ip addr in db.

As part of this change, performed following refactoring:
- Renamed GlusterUtils to GlusterDBUtils and moved it to dao module
- Modified the test case in GetAddedGlusterServersQueryTest accordingly

Change-Id: Iab25afc5cb0acb5be288c005d1110c1ad0a54b2d
Signed-off-by: Shireesh Anjal <sanjal@redhat.com>