packaging: Move dependencies to a JBoss module
Currently we package dependencies as .jar files in the lib directories
of the .ear and in the .war of RESTAPI. This has the inconvenient that
those dependencies don't have an associated artifact and also can't be
shared from different applications. This patch moves all those
dependencies to the dependencies jboss module, so that they are added to
the corresponding artifact, and can then be easily installed and shared.

Change-Id: I82a775915f4b2066ef14961cf69c2a35d0673133
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>