core: Create artifact for PostgreSQL module
Currently we are deploying the PostgreSQL module using the "setup"
profile, but we don't have an artifact for the required module. This
patch reuses the module description that we had and uses it to build a
new "dependencies" artifact that can then be easily installed.

Change-Id: Ibcb409420a724dd30bf1485a78f0c9828903fd21
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit b70f24c0ecf0eda2deeaef632d1e553158790c83)