webadmin: remove redundant variable
Remove isNew variable from StorageListModel because it's not needed.
The evaluated statement storage==null is always false,
because couple of rows before that storage name and descriptions
are evaluated, and if storage was null, then NPE would be thrown there.
Thus, storage is not null and the expression is always false, and isNew
variable is not needed.

Change-Id: Ib755a8fb39bae04349ee2617d556a589ab7cfd5c
Signed-off-by: Alissa Bonas <abonas@redhat.com>