core: Send Bonded Network with bonding
This patch locates the underlying interface
that the changed network is defined on and sends
bonding type instead of nic, if the interface is a bond.

In case we changed from a vlan network over a bond to a non-vlan
network over a bond, we need to look up the interface
using the interface name after stripping the vlan off
of it. This may be the case even if the target network
is not a vlan, since the source network was a vlan network.

Change-Id: Ifa48513b89ba7a972826d31d7fc65ee5898d5741
Bug-Url: https://bugzilla.redhat.com/909678
Signed-off-by: Muli Salem <msalem@redhat.com>