core: ImagesHandler.getAllStorageIdsForImageIds
Extracted a method to get all the relevant storage domains IDs from a
collection of DiskImages, and added some tests to it.

This method will be used in subsequent patches in the effort to decouple
ImagesHandler.PerformImagesChecks, but in the meanwhile it's main
benefit is the added unit test coverage.

Change-Id: If1b4d7231a03ea5a6d991609a65951fe4e279ef5
Signed-off-by: Allon Mureinik <amureini@redhat.com>