core: RemoveSnapshotCommand: move image validation
RemoveSnapshotCommand's canDoAction performs several validations only if
the VM has images, and several regardless.

Since ImagesHandler.PerformImagesChecks originally also validated VM
states, it was called regardless of whether the VM has images or not.
Since this is no longer the case, it can now be moved to its rightful
place.

Change-Id: I5e491f97134c3017fd53bfc159cf47f83ba74ae7
Signed-off-by: Allon Mureinik <amureini@redhat.com>