core: Database restore script complains about ...
Database restore script complains about existing entries and fails horribly

This patch fixes a bug when restoring an existing DB.
The problem was that "Drop Database" command was used while it was
executed from the active database, it caused the drop to fail since some
objects were locked and the restore  generated lots of error while
trying to create duplicate objects.
Solution is to use the built-in dropdb function instead the "Drop
Database" SQL statement.

Change-Id: I636f269e1f6094e7fc7f55ab7fddacf6a33926ba
Signed-off-by: Eli Mesika <emesika@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=910748