engine: Allow Import/Snapshot Duplicate MAC Address
This patch, allows importing/previewing a VM
with a network interface with an existing MAC address,
even though the AllowDuplicateMacAddresses flag is false.

Instead of failing the command, we simply unplug
the network interface. When user tries to plug in
the interface, we will fail the action unless the
above flag is true.

Change-Id: I269157b37426942331002c33a22fa354ba0cf39e
Bug-Url: https://bugzilla.redhat.com/873338
Signed-off-by: Muli Salem <msalem@redhat.com>