core: Remove StreamReaderCompat
This patch removes the deprecate class StreamReaderCompat and its
ancestor, TextReaderCompat.

StreamReaderCompat was only used in a single place,
GetCACertificateQuery, which was rewritten to use FileUtil and
TextReaderCompat was not used anywhere (besides both classes' unit
tests, that is).

This patch rewrites the single usage of StreamReaderCompat so it is no
longer needed, and removes the two useless classes and their unit tests.

Change-Id: Id5a072fbcdfabb3291fd3ecd83c4c722a5a3813a
Signed-off-by: Allon Mureinik <amureini@redhat.com>