core: VmRunHandler: simplify isInternal evaluation
Some of the checks in VmRunHandler are only evaluated in the following
condition:
!vm.isAutoStartup() || !runParams.getIsInternal() && vm.isAutoStartup()

Since java evaluated boolean conditions with short circuit logic, this
can be simplified.
The right hand of the || condition is only tested if
!vm.isAutoStartup() evaluates to false, which means vm.isAutoStartup()
evaluated to true - so there is no reason to test it again.

The added readability is a bonus.

Change-Id: I0aa0ce3b0ef22739014f31f4019a35b8d230aaee
Signed-off-by: Allon Mureinik <amureini@redhat.com>