core: Make connection timeout configurable
The timeout that we currently use to connect to hosts is by default very
large, as it is the same we use for VDSM operations. The default is 180
seconds but waiting that long for connections is not very reasonable, as
establishing a connection is a quick operation, specially in a high
speed local network.

This large timeout severely affects the high availability support, as it
means that we need to wait at least 6 minutes (we need at least 2
failures in a row) when a host fails and starts dropping packets before
we can declare it failed and start the recovery process.

This patch introduces two new parameters that control the time that it
takes to react when there are connection issues:

1. vdsConnectionTimeout (default is 3 minutes)

Timeout for establishment of connections with hosts. This should be
quite small, a few seconds at most, as it the TCP handshake with hosts
should be very quick in most networks.

At the moment we are using 3 minutes in order to preserve the behaviour
before this modification, but this will probably be changed to a much
smaller value (2 seconds, for example) in the future.

2. vdsRetries (default is 3)

The number of times to retry host operations when there are IO errors.
The default value is 3 to match the default value of the XML RPC library
that we use, so there is no change in behaviour.

Bug-Url: https://bugzilla.redhat.com/863211
Change-Id: I55a49ae5d655c2105c5840decec81ae712e40c32
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>