packaging: Updated getEngineVersion functionality
If we use the getEngineVersion code on upgrades from
versions where the version file does not exist, we will
experience an error with opening the version file.

This patch tries to solve it by checking the presence of
the version file. If the file doesn't exist it will use
the getRpmVersion function to fetch the real version of
the engine package from the rpm.

Change-Id: I57c62cf64d0c44f824a8ec5071f2ef50ec97854e
Signed-off-by: Alex Lourie <alourie@redhat.com>