core : increment pool master version only if there's new master
When incrementing the pool master version, it should be incremented only
when there is a new master otherwise we might not be able to use the
pool again/perform unneeded reconstructs.

Incrementing the version always can be hazardous, an example is when we
have a pool with one domain which we deactivate, master version will be
incremented and then we won't be able to perform ConnectStoragePool (as
we will have now different master version in engine compared to vdsm),
reconstruct will also fail as there are no active domains and we won't
be able to have the pool operational again.

Change-Id: Ib292802ee2b7feeb95656d492f75dde549350c48
Signed-off-by: Liron Aravot <laravot@redhat.com>