packaging: changed psql parser in engine-upgrade.getCompensations
In engine-upgrade getCompensations previously parsed the output
of psql searching for "0 rows" as flag for "no records found" and
failed on locale like Italian where it is "0 righe".
It also avoid to do the wrong thing if it contains a string like
"10 rows".
Now the parser doesn't rely on a localized string for detecting an
empty set.

Change-Id: I397d609fce6826e02b0f409e81228fe7d9d7be0a
Bug-Url: https://bugzilla.redhat.com/912426
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>