core: Clean "only in db" async tasks
When system restarts, storagePoolUp event is sent per each SPM
that becomes up.
We query the SPM and get its tasks.
If we had tasks in DB, that did not exist in SPM, we will remove them from DB.
This is an alternative to previous versions of this patch, in which I introduced
zombie treatment for these tasks.

Change-Id: Ie13ae21eaadb1bb621a7fc8ad471e60387a3699b
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>