webadmin: passing back fingerprint while updating host
While importing the hosts from a gluster cluster,
the hosts are saved along with their sshKeyFingerprint.

While updating the details of the hosts, the sshKeyFingerprint
will be passed back to the backend as it is. If it is not
passed, then backend thinks that the user is trying to update the
sshFingerprint with null and returns an update error to frontend.

Change-Id: I9e73b8bbbf5e7f2281f7a192a0157c2cc395dfc5
Bug-Url: https://bugzilla.redhat.com/888247
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>