engine: Validate Dup MAC Addresses upon Iface Plug
This patch validates that there are no plugged network interfaces
with the same MAC address as the network interface that
is being plugged.

This validation is needed, since when importing or taking a
snapshot of VM with a network interface that has a duplicate
MAC address, we unplug the network interface and do not fail
the action.

Change-Id: If1476e8bec112c0c592e5cc16deefcf67076f5dc
Bug-Url: https://bugzilla.redhat.com/873338
Signed-off-by: Muli Salem <msalem@redhat.com>